+++ pczt/src/transparent.rs
@@ -90,7 +90,6 @@ pub struct Input {
     ///   cannot produce signatures for this sighash type must not provide a signature.
     /// - Spend Finalizers must fail to finalize inputs which have signatures not matching
     ///   this sighash type.
-    #[getset(get = "pub")]
     pub(crate) sighash_type: u8,
 
     /// A map from a pubkey to the BIP 32 derivation path at which its corresponding
@@ -101,7 +100,6 @@ pub struct Input {
     /// - Individual entries may be required by a Signer.
     /// - It is not required that the map include entries for all of the used pubkeys.
     ///   In particular, it is not possible to include entries for non-BIP-32 pubkeys.
-    #[getset(get = "pub")]
     #[serde_as(as = "BTreeMap<[_; 33], _>")]
     pub(crate) bip32_derivation: BTreeMap<[u8; 33], Zip32Derivation>,
 
@@ -161,7 +159,6 @@ pub struct Output {
     /// - Individual entries may be required by a Signer.
     /// - It is not required that the map include entries for all of the used pubkeys.
     ///   In particular, it is not possible to include entries for non-BIP-32 pubkeys.
-    #[getset(get = "pub")]
     #[serde_as(as = "BTreeMap<[_; 33], _>")]
     pub(crate) bip32_derivation: BTreeMap<[u8; 33], Zip32Derivation>,